Optimizations and UI

+++ src/main/java/com/findmyrecycling/fmrenterprise/dto/Facility.java
@@ -4,13 +4,13 @@ import lombok.Data;
 
 public @Data
 class Facility {
-    private Long facilityId;
-    private Long materialId;
+    private int facilityId;
+    private int materialId;
     private String facilityName;
     private String facilityPhotos;
     private String facilityAddress;
 
-    public Facility(Long facilityId, Long materialId, String facilityName, String facilityPhotos, String facilityAddress) {
+    public Facility(int facilityId, int materialId, String facilityName, String facilityPhotos, String facilityAddress) {
         this.facilityId = facilityId;
         this.materialId = materialId;
         this.facilityName = facilityName;
